0.5v

+++ src/main/java/com/team3/DogCare/PetService/Domain/dto/FeedRequest.java
@@ -9,11 +9,11 @@ public class FeedRequest {
 
     private Long age;
 
-    private Long weight;
+    private double weight;
 
-    private Long feedCal; //사료 전체 칼로리
+    private double feedCal; //사료 전체 칼로리
 
-    private Long feedWeight; // 사료 무게
+    private double feedWeight; // 사료 무게
 
-    private Long type; //타입 (성견인지, 과체중 성견인지, 중성화or 비중성화인지)
+    private double type; //타입 (성견인지, 과체중 성견인지, 중성화or 비중성화인지)
 }
